% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pobierz_dane_kontekstowe.R
\name{pobierz_dane_kontekstowe}
\alias{pobierz_dane_kontekstowe}
\title{Pobiera dane kontekstowe o zadanym zakresie}
\usage{
pobierz_dane_kontekstowe(daneFiltr)
}
\arguments{
\item{daneFiltr}{ramka danych filtrujących - patrz opis}
}
\description{
Pobiera dane kontekstowe używane przy skalowaniach zrównujących oraz
generowaniu oszacowań umiejętności uczniów.

Zakres pobieramych obserwacji jest ograniczany na podstawie parametru
\code{daneFiltr}. Może on być dowolną ramką danych, która posiada co najmniej
jedną kolumnę wspólną z kolumnami zwracanymi przez funkcję
\code{ZPD::pobierz_dane_uczniowie_testy()}. Odfiltrowanie polega na wykonaniu
\code{semi_join(pobierz_dane_uczniowie_testy(), daneFiltr)}.
}

