% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pobierz_wyniki.R
\name{pobierz_wyniki}
\alias{pobierz_wyniki}
\title{Pobiera wyniki egzaminu z dołączonymi danymi kontekstowymi}
\usage{
pobierz_wyniki(rodzajEgzaminu, czescEgzaminu, rokEgzaminu, punktuj = TRUE,
  czyEwd = NA, idSkali = NA_integer_, skroc = TRUE)
}
\arguments{
\item{rodzajEgzaminu}{rodzaj egzaminu, ktorego wyniki maja zostac pobrane}

\item{czescEgzaminu}{czesc egzaminu, ktorego wyniki maja zostac pobrane}

\item{rokEgzaminu}{rok egzaminu, ktorego wyniki maja zostac pobrane}

\item{punktuj}{wybor, czy dane maja byc pobrane w postaci dystraktorow, czy punktow}

\item{czyEwd}{wybor, czy maja byc pobrane wyniki gromadzone przez EWD, czy ZAOU}

\item{idSkali}{identyfikator skali, ktora ma zostac zastosowana do danych}

\item{skroc}{czy do danych zastosowac skrocenia skal opisane w skali}
}
\description{
Jeśli nie zostaną podane parametry \code{czyEwd} i/lub \code{idSkali},
wtedy zastąpiopne zostaną wartościami domyślnymi.

W wypadku niepowodzenia pobrania wartości domyślnej (możliwe dla
\code{idSkali}), wyświetlony zostanie stosowny komunikat.
}

