% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/znajdz_skale_dla_testow.R
\name{znajdz_skale_dla_testow}
\alias{znajdz_skale_dla_testow}
\title{Wyszukuje skale dla wskazanych testów}
\usage{
znajdz_skale_dla_testow(idTestow, rodzajSkali = "zrównywanie")
}
\arguments{
\item{idTestow}{identyfikatory testów, dla których szukamy skali (jako wektor
lub ramka danych z kolumną \code{id_testu})}

\item{rodzajSkali}{rodzaj wyszukiwanej skali ('zrównywanie'/'ktt'/'ewd')}
}
\description{
Wyszukuje skale zadanego rodzaju (parametr \code{rodzajSkali}) dla wskazanych
testów.

Wykorzystywana np. do pobierania domyślnych skal przez funkcje
\code{pobierz_wyniki()} i \code{pobierz_zrownywanie()}.
}

